# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workerpros api",
    "description": "",
    "version": "0.1.9"
  },
  "paths": {
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "User and Job"
        ],
        "summary": "Job detail end point",
        "parameters": [
          {
            "in": "header",
            "type": "string",
            "name": "CLIENT-AUTH",
            "description": "client authorization token",
            "required": true
          },
          {
            "in": "path",
            "type": "string",
            "name": "job_id",
            "description": "job_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Request successfully executed"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Request not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/api/v1/employer/jobs/:job_id": {
      "get": {
        "tags": [
          "Employer and Job"
        ],
        "summary": "Job detail end point for employer",
        "parameters": [
          {
            "in": "path",
            "type": "string",
            "name": "job_id",
            "description": "job id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": {
              "$ref": "#/components/securitySchemes/bearerAuth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request successfully executed"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Request not found"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}